카메라 흔들림을 구현한다
카메라 흔들림이 직접적으로 카메라의 위치에 관여할 수 있도록한다.

흔들림이 끝난 뒤에는 자연스럽게 원위치로 올 수 있도록 하기 위해, 카메라의 고정된 위치를 저장하도록 한다.

덧붙여 모호한 Random함수의 호출로 인해 발생하는 스택 오버플로우를 방지하기위해, 호출하고자하는 함수에 맞추어
매개변수를 변환하도록 한다.

+++ Math.cpp
@@ -8,7 +8,7 @@ int Math::Random(int min, int max)
 
 float Math::Random(float min, float max)
 {
-	return Random(min * 1000, max * 10000) * 0.0001f;
+	return Random((int)(min * 1000), (int)(max * 10000)) * 0.0001f;
 }
 
 float Math::Random()
